homepage cleanup
